+++ jupytext/cli.py
@@ -811,7 +811,7 @@ def jupytext_single_file(nb_file, args, log):
             check_file_version(notebook, nb_file, nb_dest)
             notebook = combine_inputs_with_outputs(notebook, read(nb_dest), fmt=fmt)
         elif os.path.isfile(nb_dest):
-            action = " (destination file replaced)"
+            action = " (destination file replaced [use --update to preserve cell outputs and ids])"
         else:
             action = ""